Speed up search for "bad" instruction by using what we know to be "instructions in the loop"
The goal of part 2 is to find a single instruction which, when changed, breaks the program out of its loop. Part 1 tells us all the instructions which, when executed, are part of the infinite loop. Once we alter a candidate instruction, if any instruction next executed is part of the known loop, the modified program is also in that same infinite loop.

- The "remainder" of the loop can't be impacted by the modified instruction, since this is the first time we've seen the modified instruction.
- The first part of the loop was untouched and thus inevitably leads back to the modified instruction

It's not enough to just check that the "next" instruction isn't in the known loop:
```
jmp +2
nop +1
rest of program
```
If we convert the `jmp +2` to `acc +2`, the next-executed instruction is new, but the one after that would have been part of the infinite loop in the unmodified program.